requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
tqdm
scikit-image
matplotlib
